[ADD] <Main/components/Circle.tsx, Main.tsx> Circle 컴포넌트 만들어서 map 돌림
2020.05.20
